+++ pluslib/Support/Application.php
@@ -21,12 +21,15 @@ class Application extends Container
   public $anti_xss_header = true;
   public $login_path;
   public $assets = '/assets';
-
   public int $regenerate_session_request_counts = 100;
   public bool $session_fingerprint = true;
   public int $session_expire_time = 3600;
-
   public string $auth_class = Auth::class;
+
+  /**
+   * @var string Pluslib Version
+   */
+  const VERSION = '0.0.03';
 
   function getDefaultBindings()
   {