+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Implement Service Container
 * Cache facades by getFacadeAccessor
 * Upload Bug Fix
+* Fix Version in init.php
+* Add `version` constant to Application Class
 
 ## [v0.0.02](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.01...v0.0.02) - 8/26/2024